backend/pyproject.toml
[tool.poetry]
name = "fa-report-backend"
version = "1.0.0"
description = "FA Report Automation System — Yazaki Electronics Durango"
authors = ["Yazaki FA Lab"]

[tool.poetry.dependencies]
python          = "^3.11"
fastapi         = "^0.111.0"
uvicorn         = { extras = ["standard"], version = "^0.30.0" }
pydantic        = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlalchemy      = "^2.0.30"
asyncpg         = "^0.29.0"           # Driver async para PostgreSQL
alembic         = "^1.13.1"           # Migraciones de DB
pymupdf         = "^1.24.4"           # Motor PDF (PyMuPDF / fitz)
pillow          = "^10.3.0"           # Procesamiento de imágenes
supabase        = "^2.5.0"            # Cliente Supabase (Storage + Auth)
python-multipart = "^0.0.9"           # Upload de archivos
httpx           = "^0.27.0"           # Cliente HTTP async para tests

[tool.poetry.group.dev.dependencies]
pytest          = "^8.2.0"
pytest-asyncio  = "^0.23.7"
pytest-cov      = "^5.0.0"
black           = "^24.4.2"
ruff            = "^0.4.7"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
